pyproject.toml
[project]
name = "my-mcp-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastmcp",
    "pydantic",
    "httpx",
    "python-dotenv",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "prometheus_client"
]